Made it so if they are hit the number will go up, changed things to floats, still need to fix knockback

+++ Wrestling FMP/Assets/Scripts/GloablVariablesManager.cs
@@ -9,14 +9,14 @@ public class GloablVariablesManager : MonoBehaviour
     public int player3MashingDone;
     public int player4MashingDone;
 
-    public int minrange = 1;
-    public int maxrange = 10000;
-    public int minrangePlayer2 = 1;
-    public int maxrangePlayer2 = 10000;
-    public int minrangePlayer3 = 1;
-    public int maxrangePlayer3 = 10000;
-    public int minrangePlayer4 = 1;
-    public int maxrangePlayer4 = 10000;
+    public float minrange = 1;
+    public float maxrange = 10000;
+    public float minrangePlayer2 = 1;
+    public float maxrangePlayer2 = 10000;
+    public float minrangePlayer3 = 1;
+    public float maxrangePlayer3 = 10000;
+    public float minrangePlayer4 = 1;
+    public float maxrangePlayer4 = 10000;
 
 
     // Start is called before the first frame update
